Change return type and implementation of query function
The commit introduces changes to the 'query' function in 'Mysql.php'. It modifies the return type of the function from an array to an object. Additionally, the function implementation is updated to populate an array with retrieved rows and subsequently return them as an object. The file location has also been renamed to match case sensitivity standards.