Add main page and associated functionality
The new files include main.html in the View directory and main.js in the js directory, fully setting up the main application page. This page includes two interactive tables (wings and poles) with forms for adding new rows, as well as functionality for selecting, deleting, and moving items to storage. Relevant Vue.js methods and axios requests are also included in the main.js file.