Change MAIN_ID from variable to constant
The property MAIN_ID in the Main class of the Maps controller has been changed from being a public integer to a constant. The SQL queries in the code for the kitoro and rudak selection, which previously used this property as a variable, have been updated to reflect this change.