Refactor URL and GET parameters calculation logic
Simplified the process of calculating the URL and GET parameters in the routing.php file by delegating the responsibility to a single function, calculateUrlAndParameters, in the Request helper class. This reduces redundancy and makes the code more maintainable, as modifications on URL and GET parameters calculation need to be done at a single place.