Add 'respect' feature with Vue.js and HTML
Created a new feature named 'respect'. The Vue.js file includes methods to get data, toggle forms, add and delete rows, and perform an action if a row is selected. The HTML file is using Vue to display the data within tables on a webpage. The forms allow the user to add new data, while buttons give the options to move data to a storage or delete the selected data.