Add language switch functionality and cookie utilization
Added a function for language switching in 'farriers.js' that makes use of a newly integrated cookie to save the user's language preference. Also, adjustments were made in 'farriers.html' to display content according to the set cookie language preference. This enhances localization of site content based on the user's selected language.