Update .gitignore and add mysqli extension requirement
A new entry was added to the .gitignore file to exclude app/Database/Credentials.php, which contains sensitive data, from version control. A requirement for the mysqli PHP extension was added to composer.json. An example file for database credentials (Credentials.php.example) was also added to provide a template for developers.